task_management_system/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==45.0.4
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dominate==2.9.1
drf-nested-routers==0.94.2
executing==2.1.0
fastjsonschema==2.20.0
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
gast==0.6.0
git-filter-repo==2.47.0
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
ipython-genutils==0.2.0
itsdangerous==2.2.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
libclang==18.1.1
lxml==4.9.4
Markdown==3.7
MarkupSafe==2.1.5
mdurl==0.1.2
micawber==0.5.5
mistune==3.0.2
multimethod==1.12
mysql-connector-python==9.1.0
namex==0.0.8
narwhals==1.14.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.3
plotly==5.24.1
prettytable==3.12.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scipy==1.14.1
Send2Trash==1.8.3
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.2
tenacity==9.0.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.16.0
